new Role accessor
Now you have a list of valid role_names into the `Role::ROLES` hash.
You can ask for one of them like this `Role[:admin]`
If the Role asked for is created if not exists (I have serious doubts about the implicit create)
If the Role asded for is no defined into our hash an `ArgumentError` is raised.